chg parser, connection read part to continue reading body

+++ include/Utils.hpp
@@ -42,9 +42,9 @@ namespace utils
 	std::string					DivideStrByCRLF(std::string &data);
 	std::string					DivideStrBySpace(std::string &data);
 	void						TrimSpaceTap(std::string &trgt);
-	int							hstoi(const std::string &trgt);
-	int							ReadChunkSize(std::string &data);
-	std::string					ReadData(std::string &data, int size);
+	size_t						hstoi(const std::string &trgt);
+	size_t						ReadChunkSize(std::string &data);
+	std::string					ReadData(std::string &data, size_t size);
 	bool						CheckNameChar(std::string &data);
 	bool						CheckHostDup(std::string &header_name, std::map<std::string, std::string> &request_header);
 	void						SplitHeaderData(std::string &data, std::string &name, std::string &value);
